Add support for registering consumer at specific Message ID
Some consumers may not want to process old messages in the stream, and to
instead only start consuming from the latest. This adds a RegisterWithID
function that allows you to specify any message ID. Per the Redis documention,
`0` means the oldest message in the stream while `$` means the latest.

Fixes #6