Add ability to inject *redis.Client into producer & consumer
This adds a new field to both the `ProducerOptions` and `ConsumerOptions`
structs, `RedisClient`, that allows you to use dependency injection for the
Redis client instead of using options to create one internally.

This allows consumers to use the same Redis connection pool for their queue, and
any other operations they want to take elsewhere in their program. One use case
might be wanting a heartbeat operation that restarts the program if things seem
unhealthy.

Internally, this splits the function that creates and validates the Redis client
as being usable in to two functions. One to create the client, and another to
exercise the client to make sure it works and that Redis offers the features we
need.

Fixes #8