Upgrade to go-redis/v8
There have been a number of fixes to go-redis for Redis Cluster
compatibility. Notably, [1], which impacts the node go-redis chooses
to send a stream command to.

We're continuing to see `MOVED` errors emitted from redisqueue,
despite not overriding `MaxRedirects` when constructing our redis
client. This suggests to me that the choice of initial node may be to
blame.

This commit updates redisqueue to use go-redis/v8. In order to avoid
signature changes, we use `context.TODO()` liberally.

[1] https://github.com/go-redis/redis/issues/1501