Remove version check from Redis preflight
This commit removes the `INFO` version check from the Redis connection
preflighting.

In our automated test suite we make extensive use of
[`miniredis`](https://github.com/alicebob/miniredis) rather than
depending on a full Redis server. Miniredis is quite capable, but does
not implement the `INFO` command.

Redis 5 was released in 2018 and was a significant upgrade to Redis. I
haven't seen Redis < 5 in production since shortly after 5's
release. Therefore I believe this to be a reasonable loosening of the
constraints on redisqueue.